Tokyo Milestone (#59)
* Introduced Blueprints (#57)

* feat: trying to add Blueprint(WIP)

* fix: update import path for journal_blueprint in application.py

* fix: update import path for journal_blueprint in application.py

* enabled easy installation of packages (#56)

* requirement.txt added

* changed file name

---------


---------
