+++ api/requirements.txt
@@ -1,0 +1,5 @@
+flask
+flask_cors 
+requests 
+python-dotenv 
+groq
